requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
scipy
opencv-python
pywavelets
mlflow
matplotlib
seaborn
streamlit
pillow
tqdm
pyyaml
